Adding Emoji
The words from the edit file can be change into emoticon if you add ':' in front and at the end of the word. for example ':pizza:' will change into 🍕. Next is if you want to go to the next line of the sentences, you have to skip a line in the edit files so that the next sentence can be in the next line.